# perf(getUnhandledProps): replace lodash with vanilla js