--- a/src/lib/getUnhandledProps.js
+++ b/src/lib/getUnhandledProps.js
@@ -1,4 +1,16 @@
-import _ from 'lodash'
+const handledPropsCache = {}
+
+/**
+ * Push all `source` array elements to the `target` array if they don't already exist in `target`.
+ *
+ * @param {Array} source - An array of elements to add to the `target`
+ * @param {Array} target - An array to receive unique elements from the `source`
+ * @returns {Array} Mutated `target` array
+ */
+const pushUnique = (source, target) => source.forEach(x => {
+  if (target.indexOf(x) === -1) target.push(x)
+})
+
 /**
  * Returns an object consisting of props beyond the scope of the Component.
  * Useful for getting and spreading unknown props from the user.
@@ -7,13 +19,27 @@ import _ from 'lodash'
  * @returns {{}} A shallow copy of the prop object
  */
 const getUnhandledProps = (Component, props) => {
-  const handledProps = _.union(
-    Component.autoControlledProps,
-    _.keys(Component.defaultProps),
-    _.keys(Component.propTypes),
-  )
+  const { _meta: { name }, autoControlledProps, defaultProps, propTypes } = Component
 
-  return _.omit(props, handledProps)
+  let handledProps = handledPropsCache[name]
+
+  // ----------------------------------------
+  // Calculate handledProps once and cache
+  // ----------------------------------------
+  if (!handledProps) {
+    handledProps = []
+
+    if (autoControlledProps) pushUnique(autoControlledProps, handledProps)
+    if (defaultProps) pushUnique(Object.keys(defaultProps), handledProps)
+    if (propTypes) pushUnique(Object.keys(propTypes), handledProps)
+
+    handledPropsCache[name] = handledProps
+  }
+
+  return Object.keys(props).reduce((acc, prop) => {
+    if (handledProps.indexOf(prop) === -1) acc[prop] = props[prop]
+    return acc
+  }, {})
 }
 
 export default getUnhandledProps