@levithomason 2016-11-15
Released in `semantic-ui-react@0.60.10`.