@markerikson 2016-11-17
Hah.  I'm playing around with Semantic-UI-React on a sample project for the first time, and _just_ ran into this behavior myself (in my case, with a tabbed menu acting as parent to some panels with a couple simple forms).  After adding a table and a small form,the time to switch tabs became extremely noticeable.  A perf profile showed the issue you found here, and fortunately I ran a quick search before filing another issue :)